Procesos funcional solo falta eliminar y detalles adicionales

+++ sistema/inv/views.py
@@ -1,6 +1,8 @@
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required, permission_required
+from django.shortcuts import get_object_or_404, redirect
 from django.shortcuts import render
 from django.core.exceptions import MultipleObjectsReturned
+from django.db.models import Prefetch
 
 # Create your views here.
 import json
@@ -8,14 +10,13 @@ from django.contrib.auth.hashers import make_password
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 
-from .models import Usuario
-from .forms import UsuarioForm
+from .models import DetalleProceso, Procesos, Usuario
+from .forms import ProcesoForm, UsuarioForm
 from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 
 from .models import Inventario
 from .forms import InventarioForm
-from django.shortcuts import get_object_or_404
 from django.views.decorators.http import require_GET
 from django.db.models import Q
 from django.contrib.auth import login as auth_login, authenticate
@@ -26,10 +27,6 @@ from django.contrib import messages
 @login_required
 def inicio(request):
     return render(request, 'inicio.html')
-
-
-
-
 
 def login_view(request):
     if request.method == 'POST':
@@ -46,11 +43,11 @@ def login_view(request):
 
                     # Redirigir según el rol del usuario
                     if user.rol == 1:  # Administrador
-                        return redirect('vista_administrador')
+                        return redirect('inicio')
                     elif user.rol == 2:  # Usuario
-                        return redirect('vista_usuario')
+                        messages.error(request, "Tu cuenta es Usuario. Aun no hay vista Usuario.")
                     elif user.rol == 3:  # Invitado
-                        return redirect('vista_invitado')
+                        messages.error(request, "Tu cuenta es Invitado. Aun no hay vista Invitado.")
                 else:
                     messages.error(request, "Tu cuenta está inactiva. Contacta al administrador.")
             else:
@@ -69,32 +66,6 @@ def login_view(request):
         return render(request, 'registration/login.html')
 
     return render(request, 'registration/login.html')
-    if request.method == 'POST':
-        correo = request.POST.get('correo')
-        password = request.POST.get('password')
-        user = authenticate(request, correo=correo, password=password)
-
-        if user is not None:
-            if user.estado == 1:  # Verificar si el estado del usuario es 1 (activo)
-                auth_login(request, user)  # Iniciar sesión
-                
-                # Verificar el rol del usuario
-                if user.rol == 1:  # Rol de Administrador
-                    return redirect('inicio')  # Redirige a la vista para administradores
-                elif user.rol == 2:  # Rol de Usuario
-                    messages.error(request, "Tu cuenta es de usuario. Aun no puedes entrar.")
-                    return render(request, 'registration/login.html')  # Redirige a la vista para usuarios
-                elif user.rol == 3:  # Rol de Invitado
-                    messages.error(request, "Tu cuenta es de invitado. Aun no puedes entrar.")
-                    return render(request, 'registration/login.html') # Redirige a la vista para invitados
-            else:
-                messages.error(request, "Tu cuenta está inactiva. Contacta al administrador.")
-                return render(request, 'registration/login.html')
-        else:
-            messages.error(request, "Credenciales inválidas.")
-            return render(request, 'registration/login.html')
-
-    return render(request, 'registration/login.html')
 
 def logout_view(request):
     logout(request)  # Cierra la sesión del usuario
@@ -102,11 +73,17 @@ def logout_view(request):
 
 
 #Vistas para usuarios
-
+@login_required
 def usuarios(request):
     user = Usuario.objects.all()
     return render(request, 'usuarios/index.html', {'user': user})
 
+@login_required
+def perfil(request):
+    usuario_actual = request.user  # Usuario autenticado
+    return render(request, 'usuarios/perfil.html', {'usuario': usuario_actual})
+
+@login_required
 def crear_usuario(request):
     formulario = UsuarioForm(request.POST or None)
     if formulario.is_valid():
@@ -116,10 +93,16 @@ def crear_usuario(request):
         return redirect('usuarios')
     return render(request, 'usuarios/crear.html', {'formulario': formulario})
 
-@login_required
 def eliminar_usuario(request, id):
-    user = Usuario.objects.get(Usuario, id=id)
-    user.delete()
+    user = get_object_or_404(Usuario, id=id)
+    
+    # Verificar que no sea el propio usuario
+    if request.user != user:
+        user.delete()
+        messages.success(request, 'Usuario eliminado correctamente.')
+    else:
+        messages.error(request, 'No puedes eliminar tu propio usuario.')
+
     return redirect('usuarios')
 
 def editar_usuario(request, id):
@@ -183,7 +166,9 @@ def eliminar_inventario(request, id):
 # Vistas para los procesos
 
 def procesos(request):
-    return render(request, 'procesos/index.html')
+    procesos = Procesos.objects.all()
+    return render(request, 'procesos/index.html', {'procesos': procesos})
+    
 
 
 @require_GET
@@ -192,9 +177,125 @@ def buscar_inventario(request):
     inventario = Inventario.objects.filter(nombre__icontains=query).values('id', 'modelo')
     return JsonResponse(list(inventario), safe=False)
     
+
 def crear_procesos(request):
-    inventario = Inventario.objects.filter(estado=1)  # Filtra solo estado=3
-    return render(request, 'procesos/crear.html', {'inventario': inventario})
+    
+    if request.method == 'POST':
+        form = ProcesoForm(request.POST)
+        equipos_seleccionados = request.POST.getlist('equipos')  # Obtener equipos seleccionados
+
+        if form.is_valid():
+            # Guardar el proceso principal
+            proceso = form.save(commit=False)
+            proceso.id_autorizo = 7  # Valor fijo
+            proceso.id_autoriza_entrega = 7  # Valor fijo
+            proceso.estado = 1  # Estado inicial
+            proceso.save()
+
+            # Crear detalles del proceso para cada equipo seleccionado
+            equipos_seleccionados = request.POST.getlist('equipos')
+            for equipo_id in equipos_seleccionados:
+                equipo = Inventario.objects.get(id=equipo_id)  # Obtén la instancia
+                DetalleProceso.objects.create(
+                    proceso=proceso,
+                    inventario=equipo
+                )
+                # Actualizar estado del equipo a inactivo
+                Inventario.objects.filter(id=equipo_id).update(estado=2)
+
+            messages.success(request, 'Proceso creado exitosamente!')
+            return redirect('procesos')  # Redirigir a la lista de procesos
+        else:
+            messages.error(request, 'Por favor corrige los errores en el formulario.')
+
+    else:
+        form = ProcesoForm()
+
+    inventario = Inventario.objects.filter(estado=1)
+    return render(request, 'procesos/crear.html', {'form': form, 'inventario': inventario})
+
+def obtener_detalles_proceso(request, id):
+    try:
+        proceso = Procesos.objects.select_related(
+            'solicitante', 
+            'responsable'
+        ).prefetch_related(
+            Prefetch('detalles', queryset=DetalleProceso.objects.select_related('inventario'))
+        ).get(id=id)
+        
+        detalles_equipos = []
+        for detalle in proceso.detalles.all():
+            equipo = detalle.inventario
+            detalles_equipos.append({
+                'id_equipo': equipo.id,
+                'tipo_equipo': equipo.tipo_equipo,
+                'serie': equipo.serie,
+                'marca': equipo.marca,
+                'modelo': equipo.modelo,
+                'observaciones': equipo.observaciones
+            })
+        
+        data = {
+            'proceso': {
+                'solicitante': f"{proceso.solicitante}",
+                'responsable': f"{proceso.responsable}",
+                'fecha_regreso': proceso.fecha_regreso.strftime('%d/%m/%Y') if proceso.fecha_regreso else 'N/A',
+                'ubicacion': proceso.ubicacion,
+                'tipo': proceso.tipo,
+                'descripcion': proceso.descripcion
+            },
+            'equipos': detalles_equipos
+        }
+        return JsonResponse(data)
+    
+    except Procesos.DoesNotExist:
+        return JsonResponse({'error': 'Proceso no encontrado'}, status=404)
+    
+def quitar_equipo(request, proceso_id, equipo_id):
+    try:
+        # Obtener el detalle del proceso
+        detalle = DetalleProceso.objects.get(proceso_id=proceso_id, inventario_id=equipo_id)
+        
+        # Cambiar el estado del equipo a "1" (En inventario)
+        Inventario.objects.filter(id=equipo_id).update(estado=1)
+        
+        # Eliminar la fila del detalle del proceso
+        detalle.delete()
+        
+        return JsonResponse({'success': True})
+    
+    except DetalleProceso.DoesNotExist:
+        return JsonResponse({'error': 'Equipo no encontrado en el proceso'}, status=404)
+    
+
+def editar_proceso(request, id):
+    proceso = Procesos.objects.get(id=id)
+    
+    if request.method == 'POST':
+        form = ProcesoForm(request.POST, instance=proceso)
+        
+        if form.is_valid():
+            form.save()
+            # Agregar lógica para agregar nuevos equipos al proceso
+            equipos_seleccionados = request.POST.getlist('equipos')
+            for equipo_id in equipos_seleccionados:
+                equipo = Inventario.objects.get(id=equipo_id)
+                DetalleProceso.objects.get_or_create(
+                    proceso=proceso,
+                    inventario=equipo
+                )
+                Inventario.objects.filter(id=equipo_id).update(estado=2)
+            
+            messages.success(request, 'Proceso editado exitosamente!')
+            return redirect('procesos')
+        else:
+            messages.error(request, 'Por favor corrige los errores en el formulario.')
+    else:
+        form = ProcesoForm(instance=proceso)
+        inventario = Inventario.objects.filter(estado=1)
+    
+    return render(request, 'procesos/editar.html', {'form': form, 'inventario': inventario})
+
 
 def agregar_equipo(request, equipo_id):
     if request.method == "POST":
